requirements.txt
streamlit
langchain
langchain-community
langchain-openai
google-generativeai
psycopg2-binary
langchain-community
